update test order
