+++ tests/test_output_dir_handling.py
@@ -36,14 +36,14 @@ class Test_housekeeping_functions(unittest.TestCase):
 
     # Define function to test
 
-    @pytest.mark.run(order=4)
+    @pytest.mark.run(order=9)
     def test_output_dir_creation(self):
         """Tests function for creating output dir"""
         output_dir_handling_main.make_output_directory(
             self.output_dir, self.logger, True, True
         )
 
-    @pytest.mark.run(order=5)
+    @pytest.mark.run(order=10)
     def test_writing_df(self):
         """Tests function for writing out created dataframe"""
         output_dir_handling_main.write_out_dataframe(
